Add files via upload
This is a visual representation of the physical model I created in MySQL. It showcases the relational database for the fitness activity application. Primary keys and foreign keys are visible attributes within each table. Additionally, the crow-feet model is shown to represent whether relationships between entities are one-to-one, one-to-many, or many-many.